Refactor agents
